requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
dstack-sdk>=0.5.0
playwright>=1.40.0
cryptography>=42.0.0
python-dateutil>=2.9.0
httpx>=0.27.0
